reorganize model module to keep classes in non-init file, update imports
